Thêm user authencation

+++ main.py
@@ -164,7 +164,11 @@ def teacher_register():
 @login_required
 @app.route('/private/key/<private>')
 def private_key(private):
-    return render_template("private_key.html", private_key = private) # biến trên front-end = biến trên back-end
+    if current_user.is_authenticated:
+        pass
+    else:
+        return redirect(url_for('index'))
+    return render_template("private_key.html", private_key=private)  # biến trên front-end = biến trên back-end
 
 @app.route('login/<email>')
 def login_again(email):
@@ -176,53 +180,62 @@ def login_again(email):
 @login_required
 @app.route('/teacher/dashboard')
 def teacher_dashboard():
-    teen = float(web3.toWei(contract.functions.balanceOf(current_user.address).call(),'ether'))
-    eth = float(web3.toWei(web3.eth.getBalance(current_user.address), 'ether'))
-    class_in_level = Class.query.filter(level=current_user.level).all()
+    if current_user.is_authenticated:
+        teen = float(web3.toWei(contract.functions.balanceOf(current_user.address).call(),'ether'))
+        eth = float(web3.toWei(web3.eth.getBalance(current_user.address), 'ether'))
+        class_in_level = Class.query.filter(level=current_user.level).all()
+    else:
+        return redirect(url_for('index'))
     return render_template("teacher_dashboard.html",teen_balanced=teen,eth_balenced=eth,Class=class_in_level)
 
 @login_required
 @app.route('/class/<Class>')
 def Class_check(Class):
-    Std = Student.query.filter(Class=Class).all()
-    return render_template("class.html",student=Std)
+    if current_user.is_authenticated:
+        Std = Student.query.filter(Class=Class).all()
+    else:
+        return redirect(url_for('index'))
+    return render_template("class.html", student=Std)
 
 @login_required
 @app.route('/student/comment/<id>',methods=["GET","POST"])
 def student_comment(id):
-    std = Student.query.filter(id=int(id)).first()
-    form = st_comment()
-    fail = Fail.query.filter(std_id=int(id)).first()
-    if form.validate_on_submit():
-        atendent = form.atendent.data
-        comment = form.comment.data
-        present = form.present.data
-        private = form.priv.data
-        if "có" in atendent:
-            new_cmt = Comment(Std_name=std.student_name, comment=comment, present=str(present), by=current_user.teacher_name, std_id=std.id)
-            db.session.add(new_cmt)
-            std.present += 1
-            db.session.commit()
-            if fail:
-                pass
-            else:
-                if present > 0:
-                    return redirect("https://teenownsystem.pythonanywhere.com/"+str(present)+"/"+str(std.address)+"/"+str(private))
+    if current_user.is_authenticated:
+        std = Student.query.filter(id=int(id)).first()
+        form = st_comment()
+        fail = Fail.query.filter(std_id=int(id)).first()
+        if form.validate_on_submit():
+            atendent = form.atendent.data
+            comment = form.comment.data
+            present = form.present.data
+            private = form.priv.data
+            if "có" in atendent:
+                new_cmt = Comment(Std_name=std.student_name, comment=comment, present=str(present), by=current_user.teacher_name, std_id=std.id)
+                db.session.add(new_cmt)
+                std.present += 1
+                db.session.commit()
+                if fail:
+                    pass
                 else:
-                    return redirect(url_for('teacher_dashboard'))
-        else:
-            new_cmt = Comment(Std_name=std.student_name, comment=comment, present=str(present),by=current_user.teacher_name, std_id=std.id)
-            db.session.add(new_cmt)
-            std.absent += 1
-            db.session.commit()
-            if fail:
-                pass
+                    if present > 0:
+                        return redirect("https://teenownsystem.pythonanywhere.com/"+str(present)+"/"+str(std.address)+"/"+str(private))
+                    else:
+                        return redirect(url_for('teacher_dashboard'))
             else:
-                if present > 0:
-                    return redirect(
-                        "https://teenownsystem.pythonanywhere.com/" + str(present) + "/" + str(std.address) + "/" + str(private))
+                new_cmt = Comment(Std_name=std.student_name, comment=comment, present=str(present),by=current_user.teacher_name, std_id=std.id)
+                db.session.add(new_cmt)
+                std.absent += 1
+                db.session.commit()
+                if fail:
+                    pass
                 else:
-                    return redirect(url_for('teacher_dashboard'))
+                    if present > 0:
+                        return redirect(
+                            "https://teenownsystem.pythonanywhere.com/" + str(present) + "/" + str(std.address) + "/" + str(private))
+                    else:
+                        return redirect(url_for('teacher_dashboard'))
+    else:
+        return redirect(url_for('index'))
     return render_template('cmt.html', form=form)
 
 @login_required
@@ -266,8 +279,11 @@ def fail_id(std_id):
 @login_required
 @app.route('/student/fail/delete/<id>')
 def fail_delete(id):
-    Fail.query.filter(id=id).delete()
-    db.session.commit()
+    if current_user.is_authenticated:
+        Fail.query.filter(id=id).delete()
+        db.session.commit()
+    else:
+        return redirect(url_for('index'))
 
 @app.route('/teacher/login', methods=['GET','POST'])
 def teacher_login():
@@ -306,153 +322,200 @@ def student_login():
 @login_required
 @app.route('/add/student',methods=['GET','POST'])
 def add_student():
-    if request.method == 'POST':
-        name = request.form.get('name')
-        Class = request.form.get('class')
-        email = request.form.get('email')
-        address = request.form.get('address')
-        password = request.form.get('password')
-        cls_id = request.form.get('ID lớp')
+    if current_user.is_authenticated:
+        if request.method == 'POST':
+            name = request.form.get('name')
+            Class = request.form.get('class')
+            email = request.form.get('email')
+            address = request.form.get('address')
+            password = request.form.get('password')
+            cls_id = request.form.get('ID lớp')
 
-        new_student = Student(student_name=name, email=email, Class=Class, address=address, password=password, cls_id=cls_id)
-        db.session.add(new_student)
-        db.session.commit()
+            new_student = Student(student_name=name, email=email, Class=Class, address=address, password=password, cls_id=cls_id)
+            db.session.add(new_student)
+            db.session.commit()
 
-        return redirect(url_for('student_dashboard'))
+            return redirect(url_for('student_dashboard'))
+    else:
+        return redirect(url_for('index'))
 
     return render_template('add_student.html')
 
 @login_required
 @app.route('/delete/student/<id>')
 def delete_student(id):
-    Student.query.filter(id=int(id)).first().delete()
-    db.session.commit()
+    if current_user.is_authenticated:
+        Student.query.filter(id=int(id)).first().delete()
+        db.session.commit()
+    else:
+        return redirect(url_for('index'))
 
 @login_required
 @app.route('/change/mail/teacher/<id>',methods=['GET','POST'])
 def teacher_change_mail(id):
-    teacher = Teacher.query.filter(id=int(id)).first()
-    form = change_mail()
-    if form.validate_on_submit():
-        email = form.email.data
-        teacher.email = email
-        db.session.commit()
-        return redirect(url_for('teacher_dashboard'))
+    if current_user.is_authenticated:
+        teacher = Teacher.query.filter(id=int(id)).first()
+        form = change_mail()
+        if form.validate_on_submit():
+            email = form.email.data
+            teacher.email = email
+            db.session.commit()
+            return redirect(url_for('teacher_dashboard'))
+    else:
+        return redirect(url_for('index'))
     return render_template("change_mail.html", form=form)
 
 @login_required
 @app.route('/change/mail/student/<id>',methods=['GET','POST'])
 def student_change_mail(id):
-    std = Student.query.filter(id=int(id)).first()
-    form = change_mail()
-    if form.validate_on_submit():
-        email = form.email.data
-        std.email = email
-        db.session.commit()
-        return redirect(url_for('student_dashboard'))
+    if current_user.is_authenticated:
+        std = Student.query.filter(id=int(id)).first()
+        form = change_mail()
+        if form.validate_on_submit():
+            email = form.email.data
+            std.email = email
+            db.session.commit()
+            return redirect(url_for('student_dashboard'))
+    else:
+        return redirect(url_for('index'))
     return render_template("change_mail.html", form=form)
 
 @login_required
 @app.route('/change/password/teacher/<id>',methods=['GET','POST'])
 def teacher_change_pass(id):
-    teacher = Teacher.query.filter(id=int(id)).first()
-    form = change_password()
-    if form.validate_on_submit():
-        password = form.email.data
-        teacher.email = password
-        db.session.commit()
-        return redirect(url_for('teacher_dashboard'))
+    if current_user.is_authenticated:
+        teacher = Teacher.query.filter(id=int(id)).first()
+        form = change_password()
+        if form.validate_on_submit():
+            password = form.email.data
+            teacher.email = password
+            db.session.commit()
+            return redirect(url_for('teacher_dashboard'))
+    else:
+        return redirect(url_for('index'))
     return render_template("change_password.html", form=form)
 
 @login_required
 @app.route('/change/password/student/<id>',methods=['GET','POST'])
 def student_change_pass(id):
-    std = Student.query.filter(id=int(id)).first()
     form = change_password()
-    if form.validate_on_submit():
-        password = form.password.data
-        std.password = password
-        db.session.commit()
-        return redirect(url_for('student_dashboard'))
+    if current_user.is_authenticated:
+        std = Student.query.filter(id=int(id)).first()
+        form = change_password()
+        if form.validate_on_submit():
+            password = form.password.data
+            std.password = password
+            db.session.commit()
+            return redirect(url_for('student_dashboard'))
+        else:
+            return redirect(url_for('index'))
     return render_template("change_password.html", form=form)
 
 @login_required
 @app.route('/change/address/teacher/<id>',methods=['GET','POST'])
 def teacher_change_address(id):
-    teacher = Teacher.query.filter(id=int(id)).first()
     form = change_address()
-    if form.validate_on_submit():
-        address = form.email.data
-        teacher.address = address
-        db.session.commit()
-        return redirect(url_for('teacher_dashboard'))
+    if current_user.is_authenticated:
+        teacher = Teacher.query.filter(id=int(id)).first()
+        form = change_address()
+        if form.validate_on_submit():
+            address = form.email.data
+            teacher.address = address
+            db.session.commit()
+            return redirect(url_for('teacher_dashboard'))
+    else:
+        return redirect(url_for('index'))
     return render_template("change_address.html", form=form)
 
 @login_required
 @app.route('/change/address/student/<id>',methods=['GET','POST'])
 def student_change_address(id):
-    std = Student.query.filter(id=int(id)).first()
-    form = change_address()
-    if form.validate_on_submit():
-        address = form.address.data
-        std.address = address
-        db.session.commit()
-        return redirect(url_for('student_dashboard'))
+    if current_user.is_authenticated:
+        std = Student.query.filter(id=int(id)).first()
+        form = change_address()
+        if form.validate_on_submit():
+            address = form.address.data
+            std.address = address
+            db.session.commit()
+            return redirect(url_for('student_dashboard'))
+    else:
+        return redirect(url_for('index'))
     return render_template("change_address.html", form=form)
 
 @login_required
 @app.route('/student/homework/check')
 def student_homework_check():
-    std = current_user.cls_id
-    hw = Homework.query.filter(Cls=std).all()
-    hwk = hw[::-1]
+    if current_user.is_authenticated:
+        std = current_user.cls_id
+        hw = Homework.query.filter(Cls=std).all()
+        hwk = hw[::-1]
+    else:
+        return redirect(url_for('index'))
     return render_template("student_homework.html",post=hwk)
 
 @login_required
 @app.route('/teacher/homework/check')
 def teacher_homework_check():
-    std = current_user.teacher_name
-    hw = Homework.query.filter(by=std).all()
-    hwk = hw[::-1]
+    if current_user.is_authenticated:
+        std = current_user.teacher_name
+        hw = Homework.query.filter(by=std).all()
+        hwk = hw[::-1]
+    else:
+        return redirect(url_for('index'))
     return render_template("teacher_homework.html",post=hwk)
 
 @login_required
 @app.route('/student/comment/check')
 def student_comment_check():
-    std = current_user.id
-    hw = Comment.query.filter(std_id=std).all()
-    hwk = hw[::-1]
+    if current_user.is_authenticated:
+        std = current_user.id
+        hw = Comment.query.filter(std_id=std).all()
+        hwk = hw[::-1]
+    else:
+        return redirect(url_for('index'))
     return render_template("student_comment.html",post=hwk)
 
 @login_required
 @app.route('/teacher/comment/check')
 def teacher_comment_check():
-    std = current_user.teacher_name
-    hw = Comment.query.filter(by=std).all()
-    hwk = hw[::-1]
+    if current_user.is_authenticated:
+        std = current_user.teacher_name
+        hw = Comment.query.filter(by=std).all()
+        hwk = hw[::-1]
+    else:
+        return redirect(url_for('index'))
     return render_template("teacher_comment.html",post=hwk)
 
 @login_required
 @app.route('/student/fail/check')
 def student_fail_check():
-    std = current_user.id
-    hw = Fail.query.filter(std_id=std).all()
-    hwk = hw[::-1]
+    if current_user.is_authenticated:
+        std = current_user.id
+        hw = Fail.query.filter(std_id=std).all()
+        hwk = hw[::-1]
+    else:
+        return redirect(url_for('index'))
     return render_template("student_fail.html",post=hwk)
 
 @login_required
 @app.route('/teacher/fail/check')
 def teacher_fail_check():
-    hw = Homework.query.all()
-    hwk = hw[::-1]
+    if current_user.is_authenticated:
+        hw = Homework.query.all()
+        hwk = hw[::-1]
+    else:
+        return redirect(url_for('index'))
     return render_template("teacher_homework.html",post=hwk)
 
 @login_required
 @app.route('student/dashboard')
 def student_dashboard():
-    teen = float(web3.toWei(contract.functions.balanceOf(current_user.address).call(), 'ether'))
-    eth = float(web3.toWei(web3.eth.getBalance(current_user.address), 'ether'))
-    class_in_level = Class.query.filter(level=current_user.level).all()
+    if current_user.is_authenticated:
+        teen = float(web3.toWei(contract.functions.balanceOf(current_user.address).call(), 'ether'))
+        eth = float(web3.toWei(web3.eth.getBalance(current_user.address), 'ether'))
+        class_in_level = Class.query.filter(level=current_user.level).all()
+    else:
+        return redirect(url_for('index'))
     return render_template("student_dashboard.html", teen_balanced=teen, eth_balenced=eth, Class=class_in_level)
 
 @app.route('/')
